- First Commit : Still some changes to be done

+++ GeolocFlickr/FJReaderView.h
@@ -2,8 +2,8 @@
 //  FJReaderView.h
 //  MengaReader
 //
-//  Created by orsys on 21/05/2014.
-//  Copyright (c) 2014 François Juteau. All rights reserved.
+//  Created by François Juteau on 21/05/2014.
+//  Copyright (c) 2014 FJ. All rights reserved.
 //
 
 #import <UIKit/UIKit.h>
@@ -11,7 +11,7 @@
 @protocol FJReaderViewDelegate <NSObject>
 
 - (UIView *) readerView:(id)sender pageAtIndex:(int)index;
-- (int) numberOfPagesInReaderView:(id)readerView;
+- (NSUInteger) numberOfPagesInReaderView:(id)readerView;
 
 @end
 
@@ -19,6 +19,7 @@
 
 // La classe qui voudra être delegate devra se conformé au protocole <FRReaderViewDelegate>
 @property (weak, nonatomic) id<FJReaderViewDelegate> delegate;
+
 @property (readonly, nonatomic) int PageCounter;
 
 - (void) showPage:(int)index animated:(BOOL)isAnimated;
